// docs/types.ts
export type RateCardMetadata = RadioMetadata | TVMetadata | OOHMetadata | DIGITALMetadata;
// ...
export interface RateCardBase {
  mediaPartnerId: string;
  mediaType: 'FM' | 'TV' | 'OOH' | 'DIGITAL';
  isActive?: boolean;
  metadata: RateCardMetadata;
}

export interface RateCard extends RateCardBase {
  id: string;
  mediaPartnerName: string;
  createdAt: string;
  updatedAt: string;
}